+++ Plugin.cs
@@ -8,7 +8,7 @@ namespace ExitDoorManager;
 public class Plugin : BaseUnityPlugin {	
   public const string PLUGIN_GUID = "com.earthlingOnFire.ExitDoorManager";
   public const string PLUGIN_NAME = "Exit Door Manager";
-  public const string PLUGIN_VERSION = "1.0.2";
+  public const string PLUGIN_VERSION = "1.1.0";
 
   private void Start() {
     gameObject.hideFlags = HideFlags.HideAndDontSave;